refactor: eliminate gosec warnings through int32 type alignment (#8)
Root Cause Fix: - Configure sqlc to generate int32 for PostgreSQL
  INTEGER types - Update SQL queries with explicit ::integer casts -
  Remove unnecessary safeInt32() overflow checking function

  Type System Improvements: - Align internal types with database schema
  (int32) - Update function signatures: limit, offset, topK → int32 -
  Update config fields: MaxHistoryMessages, RAGTopK → int32 -
  Synchronize all interfaces: SessionStore, IndexerStore, SessionQuerier

  Safety Documentation: - Add #nosec comments for unavoidable API
  boundary conversions - Document why conversions are safe (validated
  ranges, bounded values) - Add validation for user input (parsedLimit
  range check 1-1000)

  Test Updates: - Fix all mock implementations to match new int32
  signatures - Update test fixtures and table-driven test data types -
  Fix environment variable test (KOOPA_ prefix)

+++ internal/rag/retriever.go
@@ -99,11 +99,12 @@ func extractQueryText(req *ai.RetrieverRequest) string {
 }
 
 // extractTopK extracts topK from request options, returns defaultK if not found
-func extractTopK(req *ai.RetrieverRequest, defaultK int) int {
+func extractTopK(req *ai.RetrieverRequest, defaultK int32) int32 {
 	if opts, ok := req.Options.(map[string]any); ok {
 		if k, exists := opts["k"]; exists {
 			if kInt, ok := k.(int); ok {
-				return kInt
+				// Safe conversion: topK values are validated to be 1-10 by callers
+				return int32(kInt) // #nosec G115 -- topK validated to be 1-10 by callers
 			}
 		}
 	}
